Fixes from review: coarse-target evaluation, mixed-precision fusion, text angle, US val split
Evaluation (resolve_eval_labels): predictions of classes a source does not
annotate were all dropped, including the correct fine class on a coarse
target: a toilet predicted on a US fixture-any object was removed while the
target was relabelled toilet, so every furniture, fixture and appliance
object on US (and CubiCasa furniture-any) counted as a miss and skewed
eval_us_PQ, the checkpoint selector. Such predictions are now kept when
they overlap (IoU > 0.5) a coarse target whose group contains their class;
other unannotated-class predictions are still dropped. Checked through the
evaluator: tp 1, fn 0; a far-away prediction is still dropped.

MSFTextFusion: the per-drawing output was index-put into a buffer of the
stage dtype, which fails under fp16/bf16 autocast (Linear outputs half).
Cast to the buffer dtype; test_textcad runs the fusion under bf16 autocast
and fails on the old code.

Text geometry and augmentation: angle sin/cos and size were computed before
augmentation, contradicting rotated or flipped geometry. Each annotation
now rides through the transforms as a short line along its reading
direction, and angle and size are recomputed afterwards
(geometry_after_transform). test_text_data checks flip and scale.

US validation: product_joint used the US test split for checkpoint
selection. multisource sources accept val_from_train {docs, seed}, holding
out whole documents from the training files; the US source holds out 3
projects, and test is only for reporting.

TextCAD schedule: 9533 / 16 with drop_last is 595 steps per epoch; decay
and evaluation steps follow.

+++ vecformer/model/vecformer/text/textcad.py
@@ -160,5 +160,5 @@ class MSFTextFusion(nn.Module):
                 att = torch.softmax(logits - penalty * dist2.unsqueeze(0), dim=-1)
                 fused = att @ v                                            # (h, Nb, d)
             fused = fused.transpose(0, 1).reshape(-1, self.attn_dim)       # (Nb, A)
-            out[ln] = self.out(fused)
+            out[ln] = self.out(fused).to(out.dtype)       # autocast: Linear output may be fp16/bf16
         return feat + out, l0_total
